refactor: fixed all GCC compiler warnings

+++ resources/Delay.h
@@ -68,7 +68,7 @@ public:
         {
             auto abIn = context.getInputBlock();
             auto abOut = context.getOutputBlock();
-            auto L = abIn.getNumSamples();
+            auto L = static_cast<int> (abIn.getNumSamples());
             auto nCh = jmin ((int) spec.numChannels, (int) abIn.getNumChannels());
 
             int startIndex, blockSize1, blockSize2;
